requirements.txt
pandas==1.5.3
datetime
plotly
matplotlib
python-bcb
sidrapy
streamlit
